Reduce code duplication in xadd code
