+++ jctools-core/src/main/java/org/jctools/queues/MpscUnboundedXaddChunk.java
@@ -15,97 +15,14 @@ package org.jctools.queues;
 
 import org.jctools.util.InternalAPI;
 
-import static org.jctools.util.UnsafeAccess.UNSAFE;
-import static org.jctools.util.UnsafeAccess.fieldOffset;
-import static org.jctools.util.UnsafeRefArrayAccess.*;
-
 @InternalAPI
-final class MpscUnboundedXaddChunk<E>
+final class MpscUnboundedXaddChunk<E> extends MpUnboundedXaddChunk<MpscUnboundedXaddChunk<E>, E>
 {
-    private static final long PREV_OFFSET = fieldOffset(MpscUnboundedXaddChunk.class, "prev");
-    private static final long NEXT_OFFSET = fieldOffset(MpscUnboundedXaddChunk.class, "next");
-    private static final long INDEX_OFFSET = fieldOffset(MpscUnboundedXaddChunk.class, "index");
-
     final static long CHUNK_CONSUMED = -1;
-
-    private final E[] buffer;
-    private final boolean pooled;
-
-    // index is a positive monotonic number, unless set to CHUNK_CONSUMED
-    private volatile long index;
-
-    private volatile MpscUnboundedXaddChunk<E> prev;
-    private volatile MpscUnboundedXaddChunk<E> next;
 
     MpscUnboundedXaddChunk(long index, MpscUnboundedXaddChunk<E> prev, int size, boolean pooled)
     {
-        buffer = allocateRefArray(size);
-        // next is null
-        spPrev(prev);
-        spIndex(index);
-        this.pooled = pooled;
-    }
+        super(index, prev, size, pooled);
 
-    boolean isPooled()
-    {
-        return pooled;
-    }
-
-    MpscUnboundedXaddChunk<E> lvNext()
-    {
-        return next;
-    }
-
-    void soNext(MpscUnboundedXaddChunk<E> value)
-    {
-        UNSAFE.putOrderedObject(this, NEXT_OFFSET, value);
-    }
-
-    MpscUnboundedXaddChunk<E> lpPrev()
-    {
-        return (MpscUnboundedXaddChunk<E>) UNSAFE.getObject(this, PREV_OFFSET);
-    }
-
-    void spPrev(MpscUnboundedXaddChunk<E> value)
-    {
-        UNSAFE.putObject(this, PREV_OFFSET, value);
-    }
-
-    long lvIndex()
-    {
-        return index;
-    }
-
-    void soIndex(long index)
-    {
-        UNSAFE.putOrderedLong(this, INDEX_OFFSET, index);
-    }
-
-    void spIndex(long index)
-    {
-        UNSAFE.putLong(this, INDEX_OFFSET, index);
-    }
-
-    void soElement(int index, E e)
-    {
-        soRefElement(buffer, calcRefElementOffset(index), e);
-    }
-
-    E lvElement(int index)
-    {
-        return lvRefElement(buffer, calcRefElementOffset(index));
-    }
-
-    E spinForElement(int index, boolean isNull)
-    {
-        E[] buffer = this.buffer;
-        long offset = calcRefElementOffset(index);
-        E e;
-        do
-        {
-            e = lvRefElement(buffer, offset);
-        }
-        while (isNull != (e == null));
-        return e;
     }
 }
